Rename files to follow kebab case.
