+++ lib/fetch-resource.js
@@ -5,8 +5,8 @@
 'use strict';
 
 const fetch = require('node-fetch');
-const processResource = require('./processResource');
-const upgradeToHttps = require('./upgradeToHttps');
+const processResource = require('./process-resource');
+const upgradeToHttps = require('./upgrade-to-https');
 
 /**
  * Download the resource
