+++ lib/generate.js
@@ -5,9 +5,9 @@
 'use strict';
 
 const sriToolbox = require('sri-toolbox');
-const guessResourceType = require('./guessResourceType');
-const getResourceTypeFromContentTypeHeader = require('./getResourceTypeFromContentTypeHeader');
-const fetchResource = require('./fetchResource');
+const guessResourceType = require('./guess-resource-type');
+const getResourceTypeFromContentTypeHeader = require('./get-resource-type-from-content-type-header');
+const fetchResource = require('./fetch-resource');
 
 /**
  * Generate SRI data for a resource
