+++ lib/get-resource-type-from-content-type-header.js
