+++ lib/guess-resource-type.js
