+++ lib/process-resource.js
